Move game logic to core The logic to actually run the game should be decoupled from backend/frontend
