+++ poker-core/Cargo.toml
@@ -9,4 +9,5 @@ edition = "2021"
 itertools = "0.10.1"
 rand = "0.8.4"
 serde = { version = "1.0.130", features = ["derive"] }
+derive_more = "0.99.17"
 #serde_json = "1.0"
