+++ poker-core/src/lib.rs
@@ -1,3 +1,6 @@
-pub mod deck;
-pub mod hand;
+pub mod cards;
+pub mod game;
 pub mod util;
+
+pub use game::table::GameInProgress;
+pub use cards::{deck, hand};
